requirements.txt
fastapi
uvicorn
python-dotenv
tableauserverclient
tableauhyperapi
psycopg2-binary
pytest
pytest-cov
dash
dash-core-components
dash-html-components
plotly
pandas
psycopg2
httpx
sqlalchemy

